Support 4K+ render resolutions

+++ renderers/webgpuraytrace/src/shaders/quad.ts
@@ -43,9 +43,9 @@ fn vert_main(@builtin(vertex_index) vertexIndex : u32) -> VertexOutput {
 
 @fragment
 fn frag_main(@builtin(position) coord: vec4<f32>) -> @location(0) vec4<f32> {
-    let x = floor(coord.x);
-    let y = floor(coord.y);
-    let index = u32(x + y * (uniforms.bufferStride)) * 4u;
+    let x = u32(floor(coord.x));
+    let y = u32(floor(coord.y));
+    let index = (x + y * u32(uniforms.bufferStride)) * 4u;
     // [0,1]
     var color = vec3<f32>(colorBuffer.data[index], colorBuffer.data[index + 1u], colorBuffer.data[index + 2u]) / uniforms.samplesPerPixel;
     // return vec4<f32>(color, 1f);
@@ -59,9 +59,9 @@ fn frag_main(@builtin(position) coord: vec4<f32>) -> @location(0) vec4<f32> {
 
 @fragment
 fn frag_normal(@builtin(position) coord: vec4<f32>) -> @location(0) vec4<f32> {
-    let x = floor(coord.x);
-    let y = floor(coord.y);
-    let index = u32(x + y * (uniforms.bufferStride)) * 4u;
+    let x = u32(floor(coord.x));
+    let y = u32(floor(coord.y));
+    let index = (x + y * u32(uniforms.bufferStride)) * 4u;
     // [0,1]
     // TODO: Convert from [-1,1] to [0,1] here instead of in the shader
     var normal = vec3<f32>(colorBuffer.data[index], colorBuffer.data[index + 1u], colorBuffer.data[index + 2u]) / uniforms.samplesPerPixel;
@@ -70,9 +70,9 @@ fn frag_normal(@builtin(position) coord: vec4<f32>) -> @location(0) vec4<f32> {
 
 @fragment
 fn frag_depth(@builtin(position) coord: vec4<f32>) -> @location(0) vec4<f32> {
-    let x = floor(coord.x);
-    let y = floor(coord.y);
-    let index = u32(x + y * (uniforms.bufferStride)) * 4u;
+    let x = u32(floor(coord.x));
+    let y = u32(floor(coord.y));
+    let index = (x + y * u32(uniforms.bufferStride)) * 4u;
     let depth = colorBuffer.data[index + 3u] / uniforms.samplesPerPixel;
     let minDepth = uniforms.minDepth;
     let maxDepth = uniforms.maxDepth;
@@ -94,9 +94,9 @@ fn frag_depth(@builtin(position) coord: vec4<f32>) -> @location(0) vec4<f32> {
 
 @fragment
 fn frag_segment(@builtin(position) coord: vec4<f32>) -> @location(0) vec4<f32> {
-    let x = floor(coord.x);
-    let y = floor(coord.y);
-    let index = u32(x + y * (uniforms.bufferStride)) * 4u;
+    let x = u32(floor(coord.x));
+    let y = u32(floor(coord.y));
+    let index = (x + y * u32(uniforms.bufferStride)) * 4u;
     var color = vec3<f32>(colorBuffer.data[index], colorBuffer.data[index + 1u], colorBuffer.data[index + 2u]) / uniforms.samplesPerPixel;
     return vec4<f32>(color, 1f);
 }
